refactor: expose the export annotation util
